+++ firestore-send-email/functions/lib/templates.js
@@ -16,13 +16,20 @@
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 const handlebars_1 = require("handlebars");
+const logs_1 = require("./logs");
 const subjHandlebars = handlebars_1.create();
 const htmlHandlebars = handlebars_1.create();
 const textHandlebars = handlebars_1.create();
 const ampHandlebars = handlebars_1.create();
-const firebase_functions_1 = require("firebase-functions");
+const attachmentsHandlebars = handlebars_1.create();
 class Templates {
     constructor(collection) {
+        this.checkTemplateExists = (name) => {
+            return this.collection
+                .where("name", "==", name)
+                .get()
+                .then((t) => !t.empty);
+        };
         this.collection = collection;
         this.collection.onSnapshot(this.updateTemplates.bind(this));
         this.templateMap = {};
@@ -54,7 +61,10 @@ class Templates {
             if (p.amp) {
                 ampHandlebars.registerPartial(p.name, p.amp);
             }
-            console.log(`registered partial '${p.name}'`);
+            if (p.attachments) {
+                logs_1.noPartialAttachmentSupport();
+            }
+            logs_1.registeredPartial(p.name);
         });
         templates.forEach((t) => {
             const tgroup = {};
@@ -70,8 +80,11 @@ class Templates {
             if (t.amp) {
                 tgroup.amp = ampHandlebars.compile(t.amp);
             }
+            if (t.attachments) {
+                tgroup.attachments = attachmentsHandlebars.compile(JSON.stringify(t.attachments), { strict: true });
+            }
             this.templateMap[t.name] = tgroup;
-            firebase_functions_1.logger.log(`loaded template '${t.name}'`);
+            logs_1.templateLoaded(t.name);
         });
         this.ready = true;
         this.waits.forEach((wait) => wait());
@@ -79,14 +92,25 @@ class Templates {
     async render(name, data) {
         await this.waitUntilReady();
         if (!this.templateMap[name]) {
-            return Promise.reject(new Error(`tried to render non-existent template '${name}'`));
+            //fallback, check if template does exist, results may be cached
+            logs_1.checkingMissingTemplate(name);
+            const templateExists = this.checkTemplateExists(name);
+            if (!templateExists)
+                return Promise.reject(new Error(`Tried to render non-existent template '${name}'`));
+            logs_1.foundMissingTemplate(name);
         }
         const t = this.templateMap[name];
+        let attachments;
+        if (t.attachments) {
+            const interpolatedAttachments = t.attachments(data);
+            attachments = JSON.parse(interpolatedAttachments);
+        }
         return {
             subject: t.subject ? t.subject(data) : null,
             html: t.html ? t.html(data) : null,
             text: t.text ? t.text(data) : null,
             amp: t.amp ? t.amp(data) : null,
+            attachments: attachments || null,
         };
     }
 }