+++ firestore-send-email/functions/src/logs.ts
@@ -80,3 +80,27 @@ export function missingUids(uids: string[]) {
     )}`
   );
 }
+
+export function noPartialAttachmentSupport() {
+  logger.warn("partial attachments are not handled and will be ignored");
+}
+
+export function registeredPartial(name: string) {
+  logger.log(`registered partial '${name}'`);
+}
+
+export function partialRegistered(name) {
+  logger.log(`registered partial '${name}'`);
+}
+
+export function templateLoaded(name) {
+  logger.log(`loaded template '${name}'`);
+}
+
+export function checkingMissingTemplate(name) {
+  logger.log(`checking missing template '${name}'`);
+}
+
+export function foundMissingTemplate(name) {
+  logger.log(`template '${name}' has been found`);
+}