+++ firestore-send-email/CHANGELOG.md
@@ -1,3 +1,39 @@
+## Version 0.1.13
+
+docs - updated POSTINSTALL docs to link to Firebase documentation
+
+## Version 0.1.12
+
+fixed - moved lease expired error message to delivery document (#726)
+
+feature - add Taiwan and Singapore Cloud Function locations (#729)
+
+fixed - empty template attachments fall back to message attachments (#719)
+
+## Version 0.1.11
+
+fixed - added fallback for finding email templates
+
+fixed - removed duplicate Warsaw location
+
+## Version 0.1.10
+
+feature - added Warsaw (europe-central2) location (#677)
+
+feature - updated Cloud Functions runtime to Node.js 14 (#660)
+
+## Version 0.1.9
+
+feature - Allow attachments within template records to be interpolated with data values passed into a new record in the mail collection. (#519)
+
+## Version 0.1.8
+
+docs - updated "SMTP connection URI" parameter description to clarify which ports are recommended for SMTP connections (#582)
+
+## Version 0.1.7
+
+fixed - Switch from `console.log` to `functions.logger.log` for cleaner log outputs.
+
 ## Version 0.1.6
 
 feature - adds support for Handlebars partials (#419)