+++ firestore-translate-text/CHANGELOG.md
@@ -1,3 +1,15 @@
+## Version 0.1.7
+
+feature - add Taiwan and Singapore Cloud Function locations (#729)
+
+## Version 0.1.6
+
+feature - added Warsaw (europe-central2) location (#677)
+
+feature - updated Cloud Functions runtime to Node.js 14 (#660)
+
+fixed - Removed code coverage check on ci
+
 ## Version 0.1.5
 
 docs - Improved documentation for input fields of type [`Map`](https://firebase.google.com/docs/firestore/manage-data/data-types#data_types).